+++ components/projectShared/Banner.js
@@ -1,62 +1,65 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { Parallax, ParallaxBanner } from 'react-scroll-parallax';
+import { Parallax, ParallaxBanner, ParallaxBannerLayer } from 'react-scroll-parallax';
 import { styled } from '@mui/material/styles';
-import { MOBILE } from "../../utils/constants";
+import { MOBILE, MOBILE_SM } from "../../utils/constants";
+import MultiFormatPhoto from "../photos/MultiFormatPhoto";
 
 const PREFIX = 'Banner';
 
 const classes = {
-  imageContainer: `${PREFIX}-imageContainer`
+  imageContainer: `${PREFIX}-imageContainer`,
+  multiFormatPhoto: `${PREFIX}-multiFormatPhoto`,
 };
 
 const Root = styled('div', {
-  shouldForwardProp: (prop) => !['heightToUse', 'imageUrl'].includes(prop)
-})(({ heightToUse, imageUrl }) => ({
+  shouldForwardProp: (prop) => !['heightToUse'].includes(prop)
+})(({ heightToUse }) => ({
   [`& .${classes.imageContainer}`]: {
     height: heightToUse,
     width: '100%',
-    background: `url(${imageUrl}) top center no-repeat`,
+    background: `top center no-repeat`,
     [`@media (max-width:${MOBILE})`]: {
       height: "40vh",
       marginBottom: '3em'
+    }
+  },
+
+  [`& .${classes.multiFormatPhoto}`]: {
+    height: 500,
+    width: '100%',
+    objectFit: 'cover',
+    [`@media (max-width:${MOBILE_SM})`]: {
+      height: 450,
+      objectPosition: 'center'
     }
   }
 }));
 
 export default function Banner(props) {
-  const heightToUse = props.height ? props.height : '50vh';
-  let content = (<ParallaxBanner
-    className={props.className}
-    layers={[
-      {
-        image: props.url,
-        amount: 0.2,
-        speed: -5
-      },
-    ]}
-    style={{
-      height: heightToUse
-    }}
-  >
-    {props.children}
-  </ParallaxBanner>)
-
-  if (props.name === 'pineyTrails') {
-    content = (<Root
-      heightToUse={heightToUse}
-      imageUrl={props.url}
-    >
-      <Parallax speed={-5} className={props.className}>
-        <div className={classes.imageContainer} />
-      </Parallax>
-    </Root>)
-  }
+  const heightToUse = props.height || '20rem';
 
   return (
-    <>
-      {content}
-    </>
+    <Root heightToUse={heightToUse}>
+      <ParallaxBanner
+        className={props.className}
+        style={{
+          height: heightToUse,
+          position: props.projectsMain ? 'absolute' : 'relative',
+          top: 0,
+          left: 0
+        }}
+      >
+        <ParallaxBannerLayer amount={0.2} speed={-5}>
+          <MultiFormatPhoto
+            alt="Main project graphic"
+            className={classes.multiFormatPhoto}
+            url={props.url}
+            photoPath={props.photoPath}
+          />
+        </ParallaxBannerLayer>
+      </ParallaxBanner>
+    </Root>
   )
 }
 
@@ -65,5 +68,7 @@ Banner.propTypes = {
   children: PropTypes.node,
   height: PropTypes.string,
   url: PropTypes.string.isRequired,
-  name: PropTypes.string
+  name: PropTypes.string,
+  photoPath: PropTypes.string.isRequired,
+  projectsMain: PropTypes.bool
 };