+++ components/projects/ProjectTemplate.js
@@ -72,7 +72,12 @@ export default function ProjectTemplate(props) {
     (<Root>
       <FadeIn>
         <div onClick={() => setLightbox_open(true)}>
-          <Banner url={photosPath + covers[name].url1} className={classes.banner} name={name} />
+          <Banner
+            url={covers[name].url1}
+            className={classes.banner}
+            name={name}
+            photoPath={photosPath}
+          />
         </div>
         {lightbox_open && (
           <Lightbox