+++ utils/colors.js
@@ -1,19 +1,13 @@
-export const greyColor = "rgb(170, 183, 196)";
 
+export const greyColor = "#4D5770";
+export const lightGreyColor = "rgb(170, 183, 196)";
 export const darkGreyColor = "rgba(0, 0, 0, 0.64)";
-
 export const greyBorderColor = "rgba(220,220,220,1)";
-
 export const shadowColor = "#e6ebf1";
-
 export const errorColor = '#ff0033';
-
 export const successColor = '#00BF8B';
-
 export const highlightColor = '#FFe359';
-
 export const orangeColor = 'rgb(255, 188, 162)';
-
 export const illustrationColors = {
   1: '#F5F5FC',
   2: '#E9EDFB',